Remove duplicate rootFSToAPIType
